升级打包脚本，增加打包web功能

+++ make-package.sh
@@ -75,6 +75,16 @@ sysnc_kernel()
 	cp -fa /lib/modules/`uname -r` $_target/lib/modules/
 }
 
+sysnc_web()
+{
+	local _target="$1"
+	local _source="$2" 
+	
+	echo "copy web ..."
+	mkdir -p $_target/var
+	cp -fa $_source $_target/var/
+}
+
 tar_pkg()
 {
 	local _target="$1"
@@ -94,6 +104,41 @@ encode_pkg()
 	rm -f $PKG_TAR
 }
 
+usage()
+{
+	echo "usage:"
+	echo "    make-package.sh [--kernel] [--web <web_dir>]"
+	echo ""
+}
+
+pack_kernel=0
+pack_web=0
+web_dir=""
+while [ ! -z "$1" ]
+do
+	case "$1" in
+		--kernel)
+		pack_kernel=1
+		shift
+		;;
+		--web)
+		pack_web=1
+		web_dir=$2
+		shift
+		;;
+		*)
+		usage
+		exit 1
+		;;
+	esac
+	shift
+done
+
+if [ $pack_web -eq 1 ] -a [ ! -d "$web_dir" ]; then
+	echo "web dir input error, please check"
+	exit 1
+fi
+
 target="/tmp/.pkg"
 rm -fr $target
 mkdir $target
@@ -103,8 +148,11 @@ rm -f $PKG_BIN
 sync_apps "$target"
 sync_rootfs "$target"
 sync_conf "$target"
-if [ "x$1" = "x--kernel" ]; then
+if [ $pack_kernel -eq 1 ]; then
 	sysnc_kernel "$target"
+fi
+if [ $pack_web -eq 1 ]; then
+	sysnc_web "$target" $web_dir
 fi
 tar_pkg "$target"
 encode_pkg
